+++ Model/PixelConfigurationInterface.php
@@ -30,6 +30,11 @@ interface PixelConfigurationInterface
     const XML_PATH_FB_PIXEL_ID = 'magestat_facebook_pixel/options/pixel_id';
 
     /**
+     * Should include tax configuration
+     */
+    const XML_PATH_FB_INC_TAX = 'magestat_facebook_pixel/settings/inc_tax';
+
+    /**
      * @return bool
      */
     public function isEnabled();
@@ -38,4 +43,9 @@ interface PixelConfigurationInterface
      * @return bool
      */
     public function getPixelId();
+
+    /**
+     * @return bool
+     */
+    public function getIncludeTax();
 }