+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+### [1.0.2](https://github.com/magestat/magento2-facebook-pixel/releases/tag/1.0.2) 2019-04-25
+
+#### Added
+- Possibility to include Tax at the product final price.
+
+#### Fixed
+- Inventory amount to be integer.
+- Removed unnecessary extra content.
+
 ### [1.0.1](https://github.com/magestat/magento2-facebook-pixel/releases/tag/1.0.1) 2019-04-10
 
 #### Fixed